fix: refactor huggingface tracer code
